requirements.txt
hydra-core==1.1.1
opencv-python==4.5.5.64
torch==1.11.0
torchvision==0.12.0
opencv-python==4.10.0.82
torch==2.3.1
torchvision==0.18.1
omegaconf==2.1.1
tqdm==4.63.1
pyquaternion==0.9.9
imgaug==0.4.0
torchmetrics==0.7.2
pytorch-lightning==1.6.2
pytorch-lightning==2.2.5
fvcore==0.1.5.post20210825
einops==0.3.2
efficientnet-pytorch==0.7.1
wandb==0.17.0